pyproject.toml
[project]
name = "dsec-ai-newsletter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "attrs==25.4.0",
    "beautifulsoup4==4.14.3",
    "cachetools==6.2.4",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "distro==1.9.0",
    "feedparser>=6.0.12",
    "google-auth==2.45.0",
    "google-genai==1.56.0",
    "greenlet==3.3.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "jsonlines==4.0.0",
    "lxml==6.0.2",
    "orjson==3.11.4",
    "playwright==1.57.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pyee==13.0.0",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "rsa==4.9.1",
    "ruff==0.14.7",
    "sniffio==1.3.1",
    "soupsieve==2.8",
    "tenacity==9.1.2",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.5.0",
    "websockets==15.0.1",
]
